+++ contest/event/acp_listener.php
@@ -17,16 +17,34 @@ use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 class acp_listener implements EventSubscriberInterface
 {
 	private \phpbb\language\language $language;
+	private \phpbb\request\request_interface $request;
+	private \phpbb\template\template $template;
+	private \phpbb\user $user;
+	private \phpbbgallery\contest\manager $contest;
 
-	public function __construct(\phpbb\language\language $language)
+	public function __construct(
+		\phpbb\language\language $language,
+		\phpbb\request\request_interface $request,
+		\phpbb\template\template $template,
+		\phpbb\user $user,
+		\phpbbgallery\contest\manager $contest
+	)
 	{
 		$this->language = $language;
+		$this->request = $request;
+		$this->template = $template;
+		$this->user = $user;
+		$this->contest = $contest;
 	}
 
 	public static function getSubscribedEvents(): array
 	{
 		return [
 			'phpbbgallery.core.acp.config.get_display_vars' => 'add_config',
+			'phpbbgallery.core.acp.albums.request_data' => 'request_album_type_data',
+			'phpbbgallery.core.acp.albums.default_data' => 'default_album_type_data',
+			'phpbbgallery.core.acp.albums.load_type_data' => 'load_album_type_data',
+			'phpbbgallery.core.acp.albums.send_to_template' => 'send_album_type_to_template',
 		];
 	}
 
@@ -64,5 +82,51 @@ class acp_listener implements EventSubscriberInterface
 
 		$return_ary['vars']['GALLERY_CONFIG'] = $settings;
 		$event['return_ary'] = $return_ary;
+	}
+
+	public function request_album_type_data(\phpbb\event\data $event): void
+	{
+		$event['album_type_data'] = [
+			'contest_start' => $this->request->variable('contest_start', ''),
+			'contest_rating' => $this->request->variable('contest_rating', ''),
+			'contest_end' => $this->request->variable('contest_end', ''),
+		];
+	}
+
+	public function default_album_type_data(\phpbb\event\data $event): void
+	{
+		$event['album_type_data'] = [
+			'contest_start' => time(),
+			'contest_rating' => 3 * 86400,
+			'contest_end' => 7 * 86400,
+		];
+	}
+
+	public function load_album_type_data(\phpbb\event\data $event): void
+	{
+		$album_data = (array) $event['album_data'];
+		if ((int) ($album_data['album_type'] ?? -1) === (int) \phpbbgallery\core\block::TYPE_CONTEST)
+		{
+			$event['album_type_data'] = $this->contest->get_contest((int) $album_data['album_id'], 'album');
+			return;
+		}
+
+		$this->default_album_type_data($event);
+	}
+
+	public function send_album_type_to_template(\phpbb\event\data $event): void
+	{
+		$this->language->add_lang('contest_acp', 'phpbbgallery/contest');
+		$album_data = (array) $event['album_data'];
+		$type_data = (array) $event['album_type_data'];
+		$start = (int) ($type_data['contest_start'] ?? time());
+		$this->template->assign_vars([
+			'S_ALBUM_ORIG_CONTEST' => (int) ($event['old_album_type'] ?? -1) === (int) \phpbbgallery\core\block::TYPE_CONTEST,
+			'S_ALBUM_CONTEST' => (int) ($album_data['album_type'] ?? -1) === (int) \phpbbgallery\core\block::TYPE_CONTEST,
+			'ALBUM_CONTEST' => (int) \phpbbgallery\core\block::TYPE_CONTEST,
+			'S_CONTEST_START' => $this->user->format_date($start, 'Y-m-d H:i'),
+			'CONTEST_RATING' => $this->user->format_date($start + (int) ($type_data['contest_rating'] ?? 0), 'Y-m-d H:i'),
+			'CONTEST_END' => $this->user->format_date($start + (int) ($type_data['contest_end'] ?? 0), 'Y-m-d H:i'),
+		]);
 	}
 }